+++ src/App/features/movies/MovieDetailsPage/MovieCover/styled.js
@@ -1,5 +1,5 @@
 import styled, { keyframes } from "styled-components";
-import { ReactComponent as StarIcon } from "../../../../images/VectorStar.svg";
+import { ReactComponent as StarIcon } from "../../../../../images/VectorStar.svg";
 
 export const MovieCoverBlackBars = styled.img`
   position: absolute;
